refactor parsers for pdp subcommands into individual .py files
