+++ diagnostic_primers/scripts/logger.py
@@ -1,0 +1,91 @@
+# -*- coding: utf-8 -*-
+"""Module providing a script logger
+
+(c) The James Hutton Institute 2017-2019
+
+Author: Leighton Pritchard
+Contact: leighton.pritchard@hutton.ac.uk
+
+Leighton Pritchard,
+Information and Computing Sciences,
+James Hutton Institute,
+Errol Road,
+Invergowrie,
+Dundee,
+DD6 9LH,
+Scotland,
+UK
+
+The MIT License
+
+Copyright (c) 2017 The James Hutton Institute
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+"""
+
+import logging
+import os
+import sys
+import time
+
+from diagnostic_primers import last_exception
+
+
+def build_logger(name, args):
+    """Return a logger for pdp"""
+    logger = logging.getLogger("{}: {}".format(name, time.asctime))
+    logger.setLevel(logging.DEBUG)
+
+    err_handler = logging.StreamHandler(sys.stderr)
+    err_formatter = logging.Formatter("%(levelname)s: %(message)s")
+    err_handler.setFormatter(err_formatter)
+
+    # Verbose output?
+    if args.verbose:
+        err_handler.setLevel(logging.INFO)
+    else:
+        err_handler.setLevel(logging.WARNING)
+    logger.addHandler(err_handler)
+
+    # If a logfile was specified, use it
+    if args.logfile is not None:
+        logdir = os.path.split(args.logfile)[:-1]
+        try:
+            if logdir[0] != "":
+                os.makedirs(
+                    os.path.join(*os.path.split(args.logfile)[:-1]), exist_ok=True
+                )
+            logstream = open(args.logfile, "w")
+        except OSError:
+            logger.error("Could not open %s for logging", args.logfile)
+            logger.error(last_exception())
+            sys.exit(1)
+
+        err_handler_file = logging.StreamHandler(logstream)
+        err_handler_file.setFormatter(err_formatter)
+        err_handler_file.setLevel(logging.INFO)
+
+        logger.addHandler(err_handler_file)
+
+    # Report arguments
+    args.cmdline = " ".join(sys.argv)
+    logger.info("Processed arguments: %s", args)
+    logger.info("command-line: %s", args.cmdline)
+
+    return logger
