+++ diagnostic_primers/scripts/subcommands/subcmd_blastscreen.py
@@ -41,11 +41,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-from diagnostic_primers import blast
 
 from tqdm import tqdm
 
-from ..tools import (
+from diagnostic_primers import blast
+from diagnostic_primers.scripts.tools import (
     create_output_directory,
     load_config_json,
     log_clines,
