+++ diagnostic_primers/scripts/subcommands/subcmd_dedupe.py
@@ -46,8 +46,7 @@ import os
 from tqdm import tqdm
 
 from diagnostic_primers import eprimer3
-
-from ..tools import load_config_json
+from diagnostic_primers.scripts.tools import load_config_json
 
 
 def ensure_path_to(fname):
